+++ qubap/qiskit/library/__init__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python3
+
+from .hamiltonians import ladder_hamiltonian
