--- qubap/qiskit/mariana/__init__.py
@@ -1,1 +1,0 @@
-#!/usr/bin/env python3
