+++ qubap/qiskit/state_efficient_ansatz/__init__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python3
+
+from state_efficient_ansatz import ansatz_constructor
